Implemented the Loan class which will be the base for loan management system.
Here we have  loan class which will keep track of all the things
related to loans for a user. It is mapped to the user using a
one-to-one relation.

We also changed the datatype for the "balance" field in account
from long to double for more precision when we involve interest
rates from the bank into the balance of user.

configured the config.xml file accordingly and also tested the
loan class implemenatation using test class.

+++ src/BankingSystem/Transaction.java
@@ -31,14 +31,14 @@ public class Transaction {
         this.transactionDescription = UtilLog.log(transactionType);
     }
     // for transactions where amount is involved like withdraw or deposit.
-    public Transaction(Enum<Constants> transactionType, Account account, Date transactionDate, long amount) {
+    public Transaction(Enum<Constants> transactionType, Account account, Date transactionDate, double amount) {
         this.transactionDate = transactionDate;
         this.account = account;
         this.transactionType = transactionType.toString();
         this.transactionDescription = UtilLog.log(transactionType, amount);
     }
     // for transactions which have other requirements
-    public Transaction(Enum<Constants> transactionType, Account account,  Date transactionDate, long amount,Object transactionDetail) {
+    public Transaction(Enum<Constants> transactionType, Account account,  Date transactionDate, double amount,Object transactionDetail) {
         this.transactionDate = transactionDate;
         this.account = account;
         this.transactionType = transactionType.toString();
